📦 Publish Package (#8)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/react-hydration-overlay@0.0.4

### Patch Changes

-   13bed88: Remove console.logs
-   13bed88: Fix ESM issues:

    -   `__dirname` not being available in Nextjs plugin
    -   `Overlay` import not having explicit `.js` extension

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>