📦 Publish Package (#11)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/react-hydration-overlay@0.0.5

### Patch Changes

-   669ff5f: - Feature: new webpack plugin:

    ```ts
    const {
      withHydrationOverlayWebpack,
    } = require("@builder.io/react-hydration-overlay/webpack");

    /** @type {import('next').NextConfig} */
    const nextConfig = {
      reactStrictMode: true,
      webpack: (config, options) => {
        config = withHydrationOverlayWebpack({
          appRootSelector: "#__next",
          isMainAppEntryPoint: (entryPointName) =>
            !options.isServer &&
(entryPointName === "pages/_app" || entryPointName === "main-app"),
        })(config);
        return config;
      },
    };

    module.exports = nextConfig;
    ```

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>