feat: Spotlight integration with hydration-overlay (#41)
Feature Addition: Spotlight Integration in Hydration-Overlay

This pull request introduces an integration of Spotlight into the
hydration-overlay module.

Key Highlights:
- Addressed all necessary conditions outlined in the discussion
[here](https://github.com/BuilderIO/hydration-overlay/issues/40#issuecomment-1915056436).
- Note: Current integration and tests may not function as expected due
to pending integration support in Spotlight. The implementation is
primed for action once the necessary changes are merged.

Next Steps:
- Track the progress of Hydration-overlay integration in Spotlight
[here](https://github.com/getsentry/spotlight/pull/343).
- Upon completion of Spotlight integration, a simple version update for
the Spotlight package in this repository is all that's needed.