📦 Publish Package (#46)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/react-hydration-overlay@0.1.0

### Minor Changes

-   1075a84: Added a Spotlight integration in hydration-overlay

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>