📦 Publish Package (#52)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/react-hydration-overlay@0.2.0

### Minor Changes

-   e0807d1: Mark @spotlightjs/spotlight as an optional peer dependency

If you enable spotlight integration, please install the dependency
explicitly
    in your project.

    ```bash
    npm install @spotlightjs/spotlight --dev
    yarn add @spotlightjs/spotlight --dev
    pnpm add @spotlightjs/spotlight --dev
    ```

-   22b63c5: Support React 18

Transition from
[react-diff-viewer-continued](https://github.com/praneshr/react-diff-viewer)
to a maintained fork
[react-diff-viewer-continued](https://github.com/aeolun/react-diff-viewer-continued)
    that supports React 18.

-   b1a2f7b: chore: update build toolchain

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>