📦 Publish Package (#54)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/react-hydration-overlay@0.3.0

### Minor Changes

-   fe782fd: Allow @spotlightjs/spotlight v2

### Patch Changes

-   05ea38d: Fix peer-dependency range

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>